// src/types/Colors.ts
interface Colors {
  Red: string;
  Yellow: string;
  Cyan: string;
  Green: string;
  Blue: string;
  Magenta: string;
  Text: string;
  SubtleText: string;
  IgnoredText: string;
  Overlay: string;
  Surface: string;
  Base: string;
  OverlayHighlight: string;
  Highlight: string;
  InactiveHighlight: string;
}

export default Colors;